preprocessing/models/model6/pyproject.toml
[tool.poetry]
name = "model6"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
geopandas = "^1.0.1"
topojson = "^1.9"
dill = "^0.3.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
